add json schema, refactor, update limit
