+++ discovery-provider/src/models/models.py
@@ -1211,6 +1211,11 @@ class UserListeningHistory(Base):
 
     user_id = Column(Integer, primary_key=True, nullable=False, index=True)
     listening_history = Column(JSONB, nullable=False, index=False)
+    # listening_history JSON schema
+    # [
+    #   {"track_id": 1, "timestamp": "2011-01-01 00:00:00"},
+    #   {"track_id": 2, "timestamp": "2012-02-02 00:00:00"}
+    # ]
 
     Index("user_id", "listening_history", unique=True)
